refactor: Move `Token` next to `Expr`
Previously `Token` was in `prqlc-parser`, and `Expr` (one of them...) was in `prqlc-ast`. This means that we can't add a `comments` field containing a `Token` to `Expr` when working on the formatting.

I'm increasingly of the view that splitting out `prqlc-ast` may not have been the correct approach, particularly given the maturity of our library, and that a simpler approach would be to collapse some of these distinctions; particularly those which don't help split up the stages of the compiler.

(While this increases the size of `prqlc-ast`, it arguably reduces the split's significance, so I think a reasonable small step to the goal above, as well as solving an immediate issue...)
