+++ Cargo.toml
@@ -48,6 +48,7 @@ consolidate-commits = true
 
 [workspace.dependencies]
 anyhow = "1.0.86"
+enum-as-inner = "0.6.0"
 insta = {version = "1.39.0", features = ["colors", "glob", "yaml"]}
 insta-cmd = "0.6.0"
 itertools = "0.12.0"
