+++ prqlc/prqlc-ast/Cargo.toml
@@ -12,7 +12,7 @@ version.workspace = true
 doctest = false
 
 [dependencies]
-enum-as-inner = "0.6.0"
+enum-as-inner = {workspace = true}
 semver = {version = "1.0.23", features = ["serde"]}
 serde = {workspace = true}
 serde_yaml = {workspace = true, optional = true}
