+++ prqlc/prqlc-ast/src/lib.rs
@@ -2,9 +2,11 @@ pub mod error;
 pub mod expr;
 mod span;
 pub mod stmt;
+pub mod token;
 mod types;
 
 pub use expr::*;
 pub use span::Span;
 pub use stmt::*;
+pub use token::*;
 pub use types::*;
