+++ prqlc/prqlc-parser/src/expr.rs
@@ -3,13 +3,13 @@ use std::collections::HashMap;
 use chumsky::prelude::*;
 
 use prqlc_ast::expr::*;
+use prqlc_ast::token::*;
 use prqlc_ast::Span;
 
 use crate::types::type_expr;
 
 use super::common::*;
 use super::interpolation;
-use super::lexer::TokenKind;
 use super::span::ParserSpan;
 
 pub fn expr_call() -> impl Parser<TokenKind, Expr, Error = PError> {
