+++ prqlc/prqlc-parser/src/lexer.rs
@@ -4,71 +4,8 @@ use chumsky::{
     text::{newline, Character},
 };
 
-use prqlc_ast::expr::*;
+use prqlc_ast::token::{Literal, Token, TokenKind, ValueAndUnit};
 use serde::{Deserialize, Serialize};
-
-#[derive(Clone, PartialEq, Serialize, Deserialize)]
-pub struct Token {
-    pub kind: TokenKind,
-    pub span: std::ops::Range<usize>,
-}
-
-#[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
-pub enum TokenKind {
-    NewLine,
-
-    Ident(String),
-    Keyword(String),
-    #[cfg_attr(
-        feature = "serde_yaml",
-        serde(with = "serde_yaml::with::singleton_map")
-    )]
-    Literal(Literal),
-    Param(String),
-
-    Range {
-        /// Whether the left side of the range is bound by the previous token
-        /// (but it's not contained in this token)
-        bind_left: bool,
-        bind_right: bool,
-    },
-    Interpolation(char, String),
-
-    /// single-char control tokens
-    Control(char),
-
-    ArrowThin,   // ->
-    ArrowFat,    // =>
-    Eq,          // ==
-    Ne,          // !=
-    Gte,         // >=
-    Lte,         // <=
-    RegexSearch, // ~=
-    And,         // &&
-    Or,          // ||
-    Coalesce,    // ??
-    DivInt,      // //
-    // Pow,         // **
-    Annotate, // @
-
-    // Aesthetics only
-    Comment(String),
-    DocComment(String),
-    /// Vec containing comments between the newline and the line wrap
-    // Currently we include the comments with the LineWrap token. This isn't
-    // ideal, but I'm not sure of an easy way of having them be separate.
-    // - The line wrap span technically includes the comments — on a newline,
-    //   we need to look ahead to _after_ the comments to see if there's a
-    //   line wrap, and exclude the newline if there is.
-    // - We can only pass one token back
-    //
-    // Alternatives:
-    // - Post-process the stream, removing the newline prior to a line wrap.
-    //   But requires a whole extra pass.
-    // - Change the functionality. But it's very nice to be able to comment
-    //   something out and have line-wraps still work.
-    LineWrap(Vec<TokenKind>),
-}
 
 /// Lex chars to tokens until the end of the input
 pub fn lexer() -> impl Parser<char, Vec<Token>, Error = Cheap<char>> {
@@ -490,96 +427,7 @@ fn end_expr() -> impl Parser<char, (), Error = Cheap<char>> {
     .rewind()
 }
 
-impl TokenKind {
-    pub fn range(bind_left: bool, bind_right: bool) -> Self {
-        TokenKind::Range {
-            bind_left,
-            bind_right,
-        }
-    }
-}
-
-// This is here because Literal::Float(f64) does not implement Hash, so we cannot simply derive it.
-// There are reasons for that, but chumsky::Error needs Hash for the TokenKind, so it can deduplicate
-// tokens in error.
-// So this hack could lead to duplicated tokens in error messages. Oh no.
-#[allow(clippy::derived_hash_with_manual_eq)]
-impl std::hash::Hash for TokenKind {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        core::mem::discriminant(self).hash(state);
-    }
-}
-
-impl std::cmp::Eq for TokenKind {}
-
-impl std::fmt::Display for TokenKind {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TokenKind::NewLine => write!(f, "new line"),
-            TokenKind::Ident(s) => {
-                if s.is_empty() {
-                    // FYI this shows up in errors
-                    write!(f, "an identifier")
-                } else {
-                    write!(f, "{s}")
-                }
-            }
-            TokenKind::Keyword(s) => write!(f, "keyword {s}"),
-            TokenKind::Literal(lit) => write!(f, "{}", lit),
-            TokenKind::Control(c) => write!(f, "{c}"),
-
-            TokenKind::ArrowThin => f.write_str("->"),
-            TokenKind::ArrowFat => f.write_str("=>"),
-            TokenKind::Eq => f.write_str("=="),
-            TokenKind::Ne => f.write_str("!="),
-            TokenKind::Gte => f.write_str(">="),
-            TokenKind::Lte => f.write_str("<="),
-            TokenKind::RegexSearch => f.write_str("~="),
-            TokenKind::And => f.write_str("&&"),
-            TokenKind::Or => f.write_str("||"),
-            TokenKind::Coalesce => f.write_str("??"),
-            TokenKind::DivInt => f.write_str("//"),
-            // TokenKind::Pow => f.write_str("**"),
-            TokenKind::Annotate => f.write_str("@{"),
-
-            TokenKind::Param(id) => write!(f, "${id}"),
-
-            TokenKind::Range {
-                bind_left,
-                bind_right,
-            } => write!(
-                f,
-                "'{}..{}'",
-                if *bind_left { "" } else { " " },
-                if *bind_right { "" } else { " " }
-            ),
-            TokenKind::Interpolation(c, s) => {
-                write!(f, "{c}\"{}\"", s)
-            }
-            TokenKind::Comment(s) => {
-                writeln!(f, "#{}", s)
-            }
-            TokenKind::DocComment(s) => {
-                writeln!(f, "#!{}", s)
-            }
-            TokenKind::LineWrap(comments) => {
-                write!(f, "\n\\ ")?;
-                for comment in comments {
-                    write!(f, "{}", comment)?;
-                }
-                Ok(())
-            }
-        }
-    }
-}
-
-impl std::fmt::Debug for Token {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}..{}: {:?}", self.span.start, self.span.end, self.kind)
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
 pub struct TokenVec(pub Vec<Token>);
 
 // impl std::fmt::Debug for TokenVec {
