+++ prqlc/prqlc-parser/src/types.rs
@@ -1,11 +1,11 @@
 use chumsky::prelude::*;
 
+use prqlc_ast::token::*;
 use prqlc_ast::*;
 
 use crate::expr::ident;
 
 use super::common::*;
-use super::lexer::TokenKind;
 
 pub fn type_expr() -> impl Parser<TokenKind, Ty, Error = PError> {
     recursive(|nested_type_expr| {
