+++ prqlc/prqlc/src/ir/pl/expr.rs
@@ -1,11 +1,12 @@
 use std::collections::HashMap;
 
 use enum_as_inner::EnumAsInner;
-
 use serde::{Deserialize, Serialize};
 
+use prqlc_ast::token::Literal;
+
 use crate::ast::generic;
-use crate::ast::{GenericTypeParam, Ident, Literal, Span, Ty};
+use crate::ast::{GenericTypeParam, Ident, Span, Ty};
 
 use crate::codegen::write_ty;
